requirements.txt
flask
line-bot-sdk
google-generativeai
gunicorn
python-dotenv
psycopg2-binary
pycryptodome
Pillow